requirements.txt
qrcode==6.1
opencv_python==3.4.6.27
Pillow==8.0.0
slackclient==2.9.2
